The type identifier is now working basedo n the 32 bit integers

+++ Blaze/Engine/Core/include/Blaze/Engine/Core/Types/TypeIdentifier.hpp
@@ -9,14 +9,14 @@ namespace blaze
 	{
 	public:
 		template<typename T>
-		static UInt64 get()
+		static UInt32 get()
 		{
 			return id<std::decay_t<T>>;
 		}
 	private:
-		inline static std::atomic<UInt64> counter{ 0 };
+		inline static std::atomic<UInt32> counter{ 0 };
 		
 		template<typename T>
-		inline static const UInt64 id = counter++;
+		inline static const UInt32 id = counter++;
 	};
 }
